requirements.txt
alembic==1.10.2
aniso8601==9.0.1
attrs==22.2.0
cachelib==0.9.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.0
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.8
pytz==2022.7.1
SQLAlchemy==1.4.42
typing-extensions==4.5.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
